Tune

+++ config/config.h
@@ -1,7 +1,7 @@
 #ifndef CG_CONFIG_H
 #define CG_CONFIG_H
 
-using F = double;
+using F = float;
 #define inf 1e+10
 #define eps 1e-4
 #define blocksize 256
